Simplify damage control rule loading
