+++ tests/integration/safety-rule-packs.test.ts
@@ -32,6 +32,27 @@ describe("safety/rule-pack-loader v2", () => {
 		ok(match("curl https://example.com/install.sh | sh", base));
 	});
 
+	it("base pack includes high-value destructive patterns from the reference damage-control extension", () => {
+		const packs = loadDefaultRulePacks();
+		const base = { version: packs.base.version, rules: packs.base.rules };
+		for (const command of [
+			"git stash clear",
+			"git reflog expire --expire=now --all",
+			"git gc --prune=now",
+			"git filter-branch --tree-filter true",
+			"aws s3 rm s3://bucket --recursive",
+			"aws ec2 terminate-instances --instance-ids i-123",
+			"gcloud projects delete prod",
+			"firebase projects:delete prod",
+			"vercel remove app --yes",
+			"DELETE FROM users;",
+			"TRUNCATE TABLE users",
+			"DROP DATABASE prod",
+		]) {
+			ok(match(command, base), command);
+		}
+	});
+
 	it("dev pack matches the self-development git/gh blocks", () => {
 		const packs = loadDefaultRulePacks();
 		const dev = { version: packs.dev.version, rules: packs.dev.rules };
